+++ packages/assets-controllers/src/NftDetectionController.ts
@@ -4,7 +4,7 @@ import {
   BaseController,
   type ControllerGetStateAction,
   type ControllerStateChangeEvent,
-} from '@metamask/base-controller/next';
+} from '@metamask/base-controller';
 import {
   toChecksumHexAddress,
   ChainId,