+++ packages/delegation-controller/src/types.ts
@@ -2,7 +2,7 @@ import type { AccountsControllerGetSelectedAccountAction } from '@metamask/accou
 import type {
   ControllerGetStateAction,
   ControllerStateChangeEvent,
-} from '@metamask/base-controller/next';
+} from '@metamask/base-controller';
 import type { KeyringControllerSignTypedMessageAction } from '@metamask/keyring-controller';
 import type { Messenger } from '@metamask/messenger';