+++ packages/json-rpc-engine/src/v2/index.ts
@@ -2,7 +2,8 @@ export { asLegacyMiddleware } from './asLegacyMiddleware';
 export { getUniqueId } from '../getUniqueId';
 export * from './JsonRpcEngineV2';
 export { JsonRpcServer } from './JsonRpcServer';
-export type { MiddlewareContext, EmptyContext } from './MiddlewareContext';
+export { MiddlewareContext } from './MiddlewareContext';
+export type { EmptyContext } from './MiddlewareContext';
 export { isNotification, isRequest, JsonRpcEngineError } from './utils';
 export type {
   Json,